refactor: relocate credential prompt utility
Move `promptForCredentials` from `utils` to `ui` for better organization. Update imports in `download.ts`, `list.ts`, and `sync.ts` to reflect the new path. This change enhances code structure by grouping UI-related functions together, improving maintainability and clarity.
